requirements.txt
# Core platform
anthropic>=0.42.0
python-dotenv>=1.0.0
rich>=13.7.0

# Web layer
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
jinja2>=3.1.4
python-multipart>=0.0.9
slowapi>=0.1.9

# Testing
pytest>=8.2.0
pytest-playwright>=0.5.0
pytest-asyncio>=0.23.0
requests>=2.32.0
